chore(models): update models
