+++ src/app/models/appointment.ts
@@ -1,19 +1,9 @@
+import { Doctor } from "./doctor";
+
 export interface Appointment {
   id: number,
   dia: string,
   horario: string,
   data_agendamento: string,
   medico: Doctor
-}
-
-export interface Doctor {
-  id: number,
-  crm: number,
-  nome: string,
-  especialidade: Specialty
-}
-
-export interface Specialty {
-  id: number;
-  nome: string;
 }
