+++ src/app/models/doctor.ts
@@ -1,0 +1,8 @@
+import { Specialty } from "./specialty";
+
+export interface Doctor {
+  id: number,
+  crm: number,
+  nome: string,
+  especialidade: Specialty
+}
