+++ src/app/models/specialty.ts
@@ -1,0 +1,4 @@
+export interface Specialty {
+  id: number;
+  nome: string;
+}
